Implemented basic PLSMesh format, not tested with erroneous data

+++ Pulsarion/src/Pulsarion/Util/Checksum.cpp
@@ -3,19 +3,19 @@
 
 namespace Pulsarion
 {
-    std::uint32_t Checksum::CRC32(const std::uint8_t* data, std::size_t size)
+    std::uint32_t Checksum::CRC32(const std::byte* data, std::size_t size)
     {
         static const std::array<uint32_t, 256> CRC32Table = GenerateCRC32Table();
 
         std::uint32_t crc = 0xFFFFFFFF;
         for (std::size_t i = 0; i < size; ++i)
         {
-            crc = (crc >> 8) ^ CRC32Table[(crc ^ data[i]) & 0xFF];
+            crc = (crc >> 8) ^ CRC32Table[(crc ^ static_cast<std::uint8_t>(data[i])) & 0xFF];
         }
         return crc ^ 0xFFFFFFFF;
     }
 
-    bool Checksum::ValidateCRC32(const std::uint8_t* data, std::size_t size, std::uint32_t checksum)
+    bool Checksum::ValidateCRC32(const std::byte* data, std::size_t size, std::uint32_t checksum)
     {
         return CRC32(data, size) == checksum;
     }
